+++ src/lib.rs
@@ -21,12 +21,6 @@ pub mod hooks;
 #[cfg(not(feature = "_nohooks"))]
 pub mod systemhooks;
 #[cfg(not(feature = "_nohooks"))]
-#[deprecated(
-    since = "0.1609.0",
-    note = "define deserialization types in client code"
-)]
-pub mod types;
-#[cfg(not(feature = "_nohooks"))]
 pub mod webhooks;
 
 #[cfg(feature = "client_api")]
@@ -40,13 +34,6 @@ pub use crate::auth::AuthError;
 pub use crate::gitlab::{
     AsyncGitlab, Gitlab, GitlabBuilder, GitlabError, ImpersonationClient, RestError,
 };
-#[cfg(not(feature = "_nohooks"))]
-#[allow(deprecated)]
-#[deprecated(
-    since = "0.1609.0",
-    note = "define deserialization types in client code"
-)]
-pub use crate::types::*;
 
 #[cfg(test)]
 mod test;